+++ CHANGELOG.md
@@ -8,6 +8,17 @@ file.
 
 ### Changed
 
+### Removed 
+
+## [0.13.3] - 2020-06-06
+### Added
+
+### Changed
+- Fix issue where doc tests could hang if stdout buffer filled (#402)
+- No longer report test failure if a `should_panic` doc test is ran
+- Clean pre-existing doc tests from target/doctests directory
+- Always print stderr output from cargo when building project
+
 ### Removed
 
 ## [0.13.2] - 2020-05-25